+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 ------------------
  * service list, service info: output json
  * request: don't fetch schema for frontend software release, speeding up slapos node instance
+ * proxy: update forwarded_partition_request table after destroying a forwarded partition
 
 1.8.3 (2022-10-17)
 ------------------